rename depending components
