+++ src/components/user-panel/component.jsx
@@ -1,6 +1,6 @@
 import {useUser, useSetUser} from "../../contexts/user-context.jsx";
-import {UserWelcome} from "./user-welcome.jsx";
-import {UserLogin} from "./user-login.jsx";
+import {UserWelcome} from "./_user-welcome.jsx";
+import {UserLogin} from "./_user-login.jsx";
 
 export function UserPanel() {
     const user = useUser();
